Split sensors.py into the sensors/ package (#168 groundwork)
Pure reorganization: base.py (SensorMetadata ABC + _common_base),
worldview.py (WorldViewMetadata), dimap.py (PleiadesMetadata), and
__init__.py (SENSORS registry, sensor_for_directory/inputs,
resolve_xml_inputs, re-exports of all public names). No behavior change;
'from asp_plot.sensors import ...' is unchanged.

Also fix the two WorldView scene-selection notebooks still calling
parser.get_id_dict()/parser.xml2poly() from before those methods moved
to the sensor readers.
